backend/requirements.txt
fastapi>=0.133.0
uvicorn[standard]>=0.41.0
supabase>=2.28.0
pydantic-settings>=2.13.1
python-dotenv>=1.2.1
openai>=2.24.0
